backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0
python-dotenv==1.0.1
ollama
psycopg2-binary
